Array contains a "lock" flag, allowing "per array" locking.

+++ postlicyd/main-postlicyd.c
@@ -192,7 +192,7 @@ static bool policy_process(server_t *pcy, const config_t *config)
 
 static int policy_run(server_t *pcy, void* vconfig)
 {
-    ssize_t search_offs = MAX(0, pcy->ibuf.len - 1);
+    ssize_t search_offs = MAX(0, (ssize_t)(pcy->ibuf.len - 1));
     int nb = buffer_read(&pcy->ibuf, pcy->fd, -1);
     const char *eoq;
     query_t  *query  = pcy->data;
